LayoutTests/fast/xpath/4XPath/Borrowed/kd_20010423-expected.txt

PASS actual is "abcabca"
PASS successfullyParsed is true

TEST COMPLETE
